Moved filtered orders controller to OrdersController
