+++ src/main/java/cat/itacademy/proyectoerp/controller/OrderController.java
@@ -4,6 +4,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -24,6 +26,7 @@ import cat.itacademy.proyectoerp.dto.CreateOrderDTO;
 import cat.itacademy.proyectoerp.dto.MessageDTO;
 import cat.itacademy.proyectoerp.dto.OrderDTO;
 import cat.itacademy.proyectoerp.dto.OrderStatusDTO;
+import cat.itacademy.proyectoerp.exceptions.ArgumentNotFoundException;
 import cat.itacademy.proyectoerp.service.OrderServiceImpl;
 
 @RestController
@@ -138,6 +141,45 @@ public class OrderController {
 		}
 		
 		return ResponseEntity.ok(messageDTO);
+	}
+	
+	//gets all the orders placed by a client
+	@GetMapping("/orders/client/{clientId}")
+	public ResponseEntity<MessageDTO> getOrdersByClientId(@PathVariable(name = "clientId") UUID id) {
+		MessageDTO output;
+		try {
+			List<Order> orders = orderService.findOrdersByClient(id);
+			output = new MessageDTO("true", "orders successfully retrieved.", orders);
+			return ResponseEntity.status(HttpStatus.OK).body(output);
+		} catch (ArgumentNotFoundException argNotFoundEx) {
+			output = new MessageDTO("False", argNotFoundEx.getMessage());
+			return ResponseEntity.status(HttpStatus.OK).body(output);
+		} catch (Exception e) {
+			output = new MessageDTO("False", e.getMessage());
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(output);
+		}
+	}
+	
+	//gets all the orders attended by an employee. Also returns the orders not assigned to any employee
+	@GetMapping("/orders/employee/{employeeId}")
+	public ResponseEntity<MessageDTO> getOrdersByEmployeeId(@PathVariable(name = "employeeId") UUID id) {
+		MessageDTO output;
+		try {
+			List<Order> employeeOrders = orderService.findOrdersByEmployeeId(id);
+			List<Order> ordersUnassigned = orderService.findOrdersByStatus(OrderStatus.UNASSIGNED);
+			List<Order> ordersJoined = Stream
+										.concat(employeeOrders.stream(), ordersUnassigned.stream())
+										.collect(Collectors.toList());
+
+			output = new MessageDTO("true", "orders successfully retrieved.", ordersJoined);
+			return ResponseEntity.status(HttpStatus.OK).body(output);
+		} catch (ArgumentNotFoundException argNotFoundEx) {
+			output = new MessageDTO("False", argNotFoundEx.getMessage());
+			return ResponseEntity.status(HttpStatus.OK).body(output);
+		} catch (Exception e) {
+			output = new MessageDTO("False", e.getMessage());
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(output);
+		}
 	}
 
 }
